MapViewer refactor
